Minor improvements - Add missing NonNull annotations for java - Change zIndex type from double to int for groundOverlays to avoid future issues on iOS platform - Documentation improvements

+++ packages/google_maps_flutter/google_maps_flutter_ios/pigeons/messages.dart
@@ -380,7 +380,7 @@ class PlatformGroundOverlay {
   final PlatformPoint? anchor;
   final double transparency;
   final double bearing;
-  final double zIndex;
+  final int zIndex;
   final bool visible;
   final bool clickable;
   final double? zoomLevel;
